requirements.txt
django==3.0.3
djangorestframework==3.11.0
dj-rest-auth==1.0.4
django-allauth==0.41.0
pyyaml==5.3
drf-yasg==1.17.1